plot mean human rating instead of correlation
We realized that it does not help to remove terms with an low
average similarity and plot a clot of correlation dependent on the 
cutting value. By this we only remove the structure that leads to any correlation. Now we still remove terms with low average similarity but plot the mean of the corresponding human ratings. We expect them to increase when removing the ones from terms with low average similarity.