add a new run which created a json representation of wikipedia articles
this script can be used on a cleaned wikipedia dump. For cleaning we use the wp2txt ruby program from (http://wp2txt.rubyforge.org/). It converts this output to a file of json objects where each line is an article. It is almost as the plain corpora we used before, just conserving the structure in paragraphs.

The script is not yes sumatra ready