tagg the oshea word definition and similarity corpus
We want to measure similarity of small corpora. To quantify the results
we want to do this on the word definitions that also oshea used in his
paper.  So, there's a dataset here (wuala folder):
Short-text-semantic-similarity-Oshea2008-TRMMUCCA20081_5  it's 80
sentences  example: 3.noon:string Noon is 12 o'clock in the middle of
the day. String is thin rope made of twisted threads, used for tying
things together or tying up parcels.  the script also tags them by using
senna